requirements.txt
crewai==0.130.0
chromadb==0.5.23
langchain-community>=0.3.0,<0.4.0
duckduckgo-search>=7.0.0,<8.0.0
pysqlite3-binary==0.5.4
boto3>=1.35.0
litellm>=1.50.0
langfuse==2.60.0
openinference-instrumentation-crewai
opentelemetry-exporter-otlp
